forms working, and DRY in donations

+++ js/logOut.js
@@ -1,5 +1,15 @@
-  // sign out
-  btnLogout.addEventListener('click', e => {
-      firebase.auth().signOut();
-      localStorage.clear();
-  })
+// sign out
+if (btnLogout) {
+    btnLogout.addEventListener('click', e => {
+        firebase.auth().signOut();
+        localStorage.clear();
+        document.location.href = '../index.html';
+    })
+}
+
+if (btnDonateAgain) {
+    document.querySelector("btnDonateAgain").addEventListener('click', e => {
+        console.log("støt")
+        document.location.href = '../pages/donation.html';
+    })
+}
